app/services/document_analysis/network_topology.py
import networkx as nx
import matplotlib.pyplot as plt
import io
from PIL import Image
import pytesseract
from typing import Dict, Any, List
from .base import DocumentAnalyzer

class NetworkTopologyAnalyzer(DocumentAnalyzer):
    """Analyzes network topology diagrams using image processing and graph analysis."""
    
    def supported_formats(self) -> List[str]:
        return ['png', 'jpg', 'jpeg']
    
    async def analyze(self, content: bytes, metadata: Dict[str, Any]) -> Dict[str, Any]:
        """
        Analyze network topology diagrams to extract system components and their relationships.
        
        Args:
            content: Raw image bytes
            metadata: Image metadata
            
        Returns:
            Dictionary containing:
                - nodes: List of system components
                - edges: List of connections between components
                - critical_paths: Identified critical paths
                - bottlenecks: Potential bottlenecks
                - redundancy_points: Points with redundancy
        """
        # Convert bytes to image
        image = Image.open(io.BytesIO(content))
        
        # Extract text from image using OCR
        text = pytesseract.image_to_string(image)
        
        # Create graph from extracted components
        G = nx.Graph()
        
        # Extract components (nodes) using OCR text analysis
        components = self._extract_components(text)
        for component in components:
            G.add_node(component['name'], **component['attributes'])
            
        # Extract connections (edges) using image analysis
        connections = self._analyze_connections(image)
        for conn in connections:
            G.add_edge(conn['from'], conn['to'], **conn['attributes'])
            
        # Analyze graph for critical paths and bottlenecks
        critical_paths = self._find_critical_paths(G)
        bottlenecks = self._identify_bottlenecks(G)
        redundancy_points = self._find_redundancy(G)
        
        return {
            "nodes": [{"name": n, "attributes": G.nodes[n]} for n in G.nodes],
            "edges": [{"from": e[0], "to": e[1], "attributes": G.edges[e]} for e in G.edges],
            "critical_paths": critical_paths,
            "bottlenecks": bottlenecks,
            "redundancy_points": redundancy_points
        }
    
    def _extract_components(self, text: str) -> List[Dict[str, Any]]:
        """Extract system components from OCR text."""
        components = []
        # TODO: Implement component extraction from OCR text
        # This would identify service names, databases, load balancers, etc.
        return components
    
    def _analyze_connections(self, image: Image) -> List[Dict[str, Any]]:
        """Analyze image to identify connections between components."""
        connections = []
        # TODO: Implement connection analysis using image processing
        # This would identify lines connecting components
        return connections
    
    def _find_critical_paths(self, G: nx.Graph) -> List[Dict[str, Any]]:
        """Find critical paths in the system topology."""
        critical_paths = []
        
        # Find all paths between entry points and critical services
        entry_points = [n for n in G.nodes if G.nodes[n].get('type') == 'entry_point']
        critical_services = [n for n in G.nodes if G.nodes[n].get('critical') == True]
        
        for start in entry_points:
            for end in critical_services:
                try:
                    paths = list(nx.all_shortest_paths(G, start, end))
                    for path in paths:
                        critical_paths.append({
                            "path": path,
                            "length": len(path) - 1,
                            "bottleneck_risk": self._calculate_path_risk(G, path)
                        })
                except nx.NetworkXNoPath:
                    continue
                    
        return critical_paths
    
    def _identify_bottlenecks(self, G: nx.Graph) -> List[Dict[str, Any]]:
        """Identify potential bottlenecks in the system."""
        bottlenecks = []
        
        # Calculate betweenness centrality
        centrality = nx.betweenness_centrality(G)
        
        # Nodes with high centrality are potential bottlenecks
        threshold = 0.5  # Configurable threshold
        for node, score in centrality.items():
            if score > threshold:
                bottlenecks.append({
                    "component": node,
                    "centrality_score": score,
                    "connected_services": list(G.neighbors(node))
                })
                
        return bottlenecks
    
    def _find_redundancy(self, G: nx.Graph) -> List[Dict[str, Any]]:
        """Identify points with redundancy in the system."""
        redundancy_points = []
        
        # Find components with multiple paths between them
        for component in G.nodes:
            neighbors = list(G.neighbors(component))
            for n1 in range(len(neighbors)):
                for n2 in range(n1 + 1, len(neighbors)):
                    paths = list(nx.all_simple_paths(G, neighbors[n1], neighbors[n2]))
                    if len(paths) > 1:
                        redundancy_points.append({
                            "component": component,
                            "connected_components": [neighbors[n1], neighbors[n2]],
                            "num_paths": len(paths)
                        })
                        
        return redundancy_points
    
    def _calculate_path_risk(self, G: nx.Graph, path: List[str]) -> float:
        """Calculate risk score for a path based on component characteristics."""
        risk_score = 0.0
        
        for i in range(len(path) - 1):
            current = path[i]
            next_node = path[i + 1]
            
            # Add node risk
            risk_score += G.nodes[current].get('risk_score', 0.0)
            
            # Add edge risk
            edge_data = G.edges[(current, next_node)]
            risk_score += edge_data.get('risk_score', 0.0)
            
        return risk_score / len(path)
