requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic<2.5,>=1.10.9  # Fixed version for guardrails-ai compatibility
chromadb>=0.4.22
langchain>=0.0.350
langgraph>=0.0.10
python-multipart>=0.0.6
Pillow>=10.0.0
pytesseract>=0.3.10
networkx>=3.1
matplotlib>=3.7.1
pydot>=1.4.2
python-jose[cryptography]>=3.3.0
langchain-community>=0.0.10
langchain-core>=0.1.10
guardrails-ai==0.3.0
llama-index>=0.9.8
llama-index-vector-stores-chroma>=0.1.3
elasticsearch>=8.11.0
llama-index-embeddings-huggingface>=0.1.3
llama-index-vector-stores-elasticsearch>=0.1.3
boto3>=1.34.0
python-magic>=0.4.27
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
moto>=4.2.0
httpx>=0.26.0
ollama>=0.1.6
kubernetes>=29.0.0  # For Kubernetes integration
docker>=7.0.0  # For Docker integration
prometheus-client>=0.19.0  # For metrics collection
